+++ apps/dashboard/app/(app)/ratelimits/[namespaceId]/logs/components/table/log-details/index.tsx
