+++ apps/dashboard/lib/trpc/routers/ratelimit/query-logs/utils.ts
@@ -1,4 +1,4 @@
-import type { RatelimitQueryLogsPayload } from "@/app/(app)/ratelimits/[namespaceId]/logs-v2/components/table/query-logs.schema";
+import type { RatelimitQueryLogsPayload } from "@/app/(app)/ratelimits/[namespaceId]/logs/components/table/query-logs.schema";
 import { getTimestampFromRelative } from "@/lib/utils";
 import type { RatelimitLogsParams } from "@unkey/clickhouse/src/ratelimits";
